refactor: 타입별 피드 컴포넌트 네이밍 통일
