+++ src/components/Feed/CollectFeed/CollectFeed.styled.ts
